Moved QueryKeyNameAttribute to a new file
